6 | Add Logging, Auditing, and propert Exception Handling

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/repository/PersonRepository.java
@@ -3,6 +3,8 @@ package com.zuehlke.securesoftwaredevelopment.repository;
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
 import com.zuehlke.securesoftwaredevelopment.config.Entity;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
+import com.zuehlke.securesoftwaredevelopment.exception.InternalServerError;
+import com.zuehlke.securesoftwaredevelopment.exception.InvalidArgumentException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
@@ -25,6 +27,8 @@ public class PersonRepository {
     }
 
     public List<Person> getAll() {
+        LOG.debug("Getting all people");
+
         List<Person> personList = new ArrayList<>();
         String query = "SELECT id, firstName, lastName, email FROM persons";
         try (Connection connection = dataSource.getConnection();
@@ -34,12 +38,16 @@ public class PersonRepository {
                 personList.add(createPersonFromResultSet(rs));
             }
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.error("Failed to get all people: " + e.getMessage());
+            throw new InternalServerError();
         }
+
         return personList;
     }
 
     public List<Person> search(String searchTerm) throws SQLException {
+        LOG.debug("Searching for people with: " + searchTerm);
+
         List<Person> personList = new ArrayList<>();
         String query = "SELECT id, firstName, lastName, email FROM persons WHERE UPPER(firstName) like UPPER('%" + searchTerm + "%')" +
                 " OR UPPER(lastName) like UPPER('%" + searchTerm + "%')";
@@ -50,10 +58,17 @@ public class PersonRepository {
                 personList.add(createPersonFromResultSet(rs));
             }
         }
+        catch (SQLException e) {
+            LOG.error("Failed to search for people: " + e.getMessage());
+            throw new InternalServerError();
+        }
+
         return personList;
     }
 
     public Person get(String personId) {
+        LOG.debug("Getting person: " + personId);
+
         String query = "SELECT id, firstName, lastName, email FROM persons WHERE id = " + personId;
         try (Connection connection = dataSource.getConnection();
              Statement statement = connection.createStatement();
@@ -62,20 +77,28 @@ public class PersonRepository {
                 return createPersonFromResultSet(rs);
             }
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.warn("Failed to get the person: " + e.getMessage());
         }
 
+        LOG.warn("Person with id " + personId + " not found");
         return null;
     }
 
     public void delete(int personId) {
+        LOG.info("Deleting person " + personId);
+
+        Person personFromDb = get("" + personId);
+        if (personFromDb == null) throw new InvalidArgumentException();
+
         String query = "DELETE FROM persons WHERE id = " + personId;
         try (Connection connection = dataSource.getConnection();
              Statement statement = connection.createStatement();
         ) {
             statement.executeUpdate(query);
+            auditLogger.audit("Successfully deleted person " + personFromDb);
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.warn("Failed to delete person: " + e.getMessage());
+            throw new InvalidArgumentException();
         }
     }
 
@@ -88,7 +111,11 @@ public class PersonRepository {
     }
 
     public void update(Person personUpdate) {
+        LOG.info("Updating person " + personUpdate.getId());
+
         Person personFromDb = get(personUpdate.getId());
+        if (personFromDb == null) throw new InvalidArgumentException();
+
         String query = "UPDATE persons SET firstName = ?, lastName = '" + personUpdate.getLastName() + "', email = ? where id = " + personUpdate.getId();
 
         try (Connection connection = dataSource.getConnection();
@@ -99,8 +126,10 @@ public class PersonRepository {
             statement.setString(1, firstName);
             statement.setString(2, email);
             statement.executeUpdate();
+            auditLogger.audit(new Entity("Person", personUpdate.getId(), personFromDb.toString(), personUpdate.toString()));
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.warn("Failed to update person: " + e.getMessage());
+            throw new InvalidArgumentException();
         }
     }
 }
