+++ src/main/java/com/zuehlke/securesoftwaredevelopment/repository/UserRepository.java
@@ -1,6 +1,10 @@
 package com.zuehlke.securesoftwaredevelopment.repository;
 
+import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
+import com.zuehlke.securesoftwaredevelopment.domain.Person;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
+import com.zuehlke.securesoftwaredevelopment.exception.InternalServerError;
+import com.zuehlke.securesoftwaredevelopment.exception.InvalidArgumentException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
@@ -15,6 +19,7 @@ import java.sql.Statement;
 public class UserRepository {
 
     private static final Logger LOG = LoggerFactory.getLogger(UserRepository.class);
+    private static final AuditLogger auditLogger = AuditLogger.getAuditLogger(UserRepository.class);
 
     private DataSource dataSource;
 
@@ -23,7 +28,18 @@ public class UserRepository {
     }
 
     public User findUser(String username) {
+        LOG.debug("Getting user with username" + username);
         String query = "SELECT id, username, password FROM users WHERE username='" + username + "'";
+        return findUserWithQuery(query);
+    }
+
+    private User findUserById(int userId) {
+        LOG.debug("Getting user with id" + userId);
+        String query = "SELECT id, username, password FROM users WHERE id='" + userId + "'";
+        return findUserWithQuery(query);
+    }
+
+    private User findUserWithQuery(String query) {
         try (Connection connection = dataSource.getConnection();
              Statement statement = connection.createStatement();
              ResultSet rs = statement.executeQuery(query)) {
@@ -34,8 +50,10 @@ public class UserRepository {
                 return new User(id, username1, password);
             }
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.error("Failed to get user: " + e.getMessage());
+            throw new InternalServerError();
         }
+
         return null;
     }
 
@@ -46,19 +64,26 @@ public class UserRepository {
              ResultSet rs = statement.executeQuery(query)) {
             return rs.next();
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.error("Failed to get user: " + e.getMessage());
+            throw new InternalServerError();
         }
-        return false;
     }
 
     public void delete(int userId) {
+        LOG.info("Deleting user " + userId);
+
+        User userFromDb = findUserById(userId);
+        if (userFromDb == null) throw new InvalidArgumentException();
+
         String query = "DELETE FROM users WHERE id = " + userId;
         try (Connection connection = dataSource.getConnection();
              Statement statement = connection.createStatement();
         ) {
             statement.executeUpdate(query);
+            auditLogger.audit("Successfully deleted user " + userFromDb);
         } catch (SQLException e) {
-            e.printStackTrace();
+            LOG.warn("Failed to delete user: " + e.getMessage());
+            throw new InvalidArgumentException();
         }
     }
 }
